+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Extra Icons Change Log
 
+### 0.4 (WIP)
+* reworked Maven wrapper `mvnw`, Gradle wrapper `gradlew` files.
+* support `README`, `CHANGELOG`, `CHANGES`, `LICENSE`, `COPYING`, `CONTRIBUTING`, `AUTHORS` files (with `.md`, `.adoc`, `.txt` or no extension).
+
 ### 0.3 (2018/08/25)
 * support Windows script `*.cmd`, `*.bat`, Powershell `*.ps1`, Bash `*.sh`, Maven wrapper `mvnw`, Gradle wrapper `gradlew` files.